#75 seperate kybot-client to another project

+++ build/script/compress.sh
@@ -21,9 +21,6 @@ cp -rf commit_SHA1 lib kybot tomcat ${package_name}/
 cp -rf conf/* ${package_name}/conf/
 cp -rf bin/* ${package_name}/bin/
 
-# add kybot files
-cp deploy/kybot.sh ${package_name}/kybot/
-
 rm -rf lib tomcat commit_SHA1
 find ${package_name} -type d -exec chmod 755 {} \;
 find ${package_name} -type f -exec chmod 644 {} \;
@@ -32,20 +29,23 @@ mkdir -p ../dist
 tar -cvzf ../dist/${package_name}.tar.gz ${package_name}
 rm -rf ${package_name}
 
-# package obf tar
 cd ../dist
-tar xzf ${package_name}.tar.gz
+# package obf tar
+if [ "$SKIP_OBF" != "1" ]; then
+    tar -xzf ${package_name}.tar.gz
 
-mv ../tmp/kylin.war ${package_name}/tomcat/webapps/kylin.war
-mv ../tmp/kylin-coprocessor-kap-${release_version}-obf.jar ${package_name}/lib/kylin-coprocessor-kap-${release_version}.jar
-mv ../tmp/kylin-storage-parquet-kap-${release_version}-obf.jar ${package_name}/lib/kylin-storage-parquet-kap-${release_version}.jar
-mv ../tmp/kylin-job-kap-${release_version}-obf.jar ${package_name}/lib/kylin-job-kap-${release_version}.jar
-mv ../tmp/kybot-client-${release_version}-obf.jar ${package_name}/kybot/kybot-client-${release_version}.jar
-tar -cvzf ${package_name}-obf.tar.gz ${package_name}
+    mv ../tmp/kylin.war ${package_name}/tomcat/webapps/kylin.war
+    mv ../tmp/kylin-coprocessor-kap-${release_version}-obf.jar ${package_name}/lib/kylin-coprocessor-kap-${release_version}.jar
+    mv ../tmp/kylin-storage-parquet-kap-${release_version}-obf.jar ${package_name}/lib/kylin-storage-parquet-kap-${release_version}.jar
+    mv ../tmp/kylin-job-kap-${release_version}-obf.jar ${package_name}/lib/kylin-job-kap-${release_version}.jar
+    mv ../tmp/kybot-client-${release_version}-obf.jar ${package_name}/kybot/kybot-client-${release_version}.jar
+    tar -cvzf ${package_name}-obf.tar.gz ${package_name}
 
-rm -r ../tmp
-rm -rf ${package_name}
+    rm -r ../tmp
+    rm -rf ${package_name}
 
-mv ../server_mapping.txt ${package_name}-obf.mapping
+    mv ../server_mapping.txt ${package_name}-obf.mapping
+fi
+
 echo "Package ready."
 ls ${package_name}*.tar.gz
